+++ crates/ruma-client-api/CHANGELOG.md
@@ -36,6 +36,8 @@ Improvements:
 - Add support for recursion on the `get_relating_events` endpoints, according to
   MSC3981 / Matrix 1.10
 - Add server support discovery endpoint, according to MSC1929 / Matrix 1.10
+- Add `dir` `Request` field on the `get_relating_events_with_rel_types` and
+  `get_relating_events_with_rel_type_and_event_type` endpoints
 
 # 0.17.4